Add -entry commandline option to override the name of the entry procedure.
This is necessary in order to use SDL 1.2 or SDL 2.0 on at least Android, OSX,
iOS, and winRT, and is optional on Windows. SDL_main.h #defines main to SDL_main,
and declares the real main() itself. Under FB that can't be done, so the solution
is to pass -entry SDL_main when compiling the main module.

No guarentees that that is all that is needed to properly initialise a
FB program under SDL.

Note: I wasn't sure where to put this global. env.clopt.entry would seem to
be the natural place, but env.clopt contains only integral datatypes,
and file-scope for other structures like fbc seems to be strictly enforced.
So it went into fbc and then passed to fbInit and put in env.