WIP Show a warning when a wstring can't be encoded for the target, when crosscompiling
This can happen if the target has a different sizeof(wstring).
Replace unconvertable characters with '?' instead of garbage.
(FIXME: this means that WSTR in the rtlib behaves differently
to the compile-time implementation.)
Converting from 4-byte to 2-byte wstring ought to performed by
emitting UTF-16 surrogate pairs instead.

Unfortunately, by the time string literals are emitted we no longer
know what line they were on. So instead print the literal in the warning
message. I had to extend errReportWarnEx to take a wstring ptr arg
to allow that, and it's not very helpful when showing a single
character result from WCHR.